feat: 크루 조회시 발자취 평균값 보내기 수정

+++ src/main/java/com/follow_me/running_mate/domain/crew/dto/response/CrewResponse.java
@@ -59,7 +59,7 @@ public class CrewResponse {
         private Integer memberCount;
         private String shortDescription;
         private String profileImageUrl;
-        private Integer footprintaverage;
+        private Long footprintaverage;
         //TODO: 크루 발자국 어떻게 처리할지 고민하기
     }
     @Getter
